# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "0.1"
    },
    "paths": {
        "/api/blog": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Obtener todas las entradas del blog",
                "operationId": "ea1c9e4eb5fdd360cfe59cdefce55814",
                "responses": {
                    "200": {
                        "description": "Lista de entradas del blog"
                    }
                }
            },
            "post": {
                "tags": [
                    "Blog"
                ],
                "summary": "Crear una nueva entrada del blog",
                "operationId": "5600c727b63d95db2095811a8e8cfa58",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Entrada del blog creada exitosamente"
                    }
                }
            }
        },
        "/api/blog/{id}": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Obtener una entrada del blog por ID",
                "operationId": "6ac9d2eb1f1ccbc64de4ae3a2bb02941",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la entrada del blog",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entrada del blog encontrada"
                    },
                    "404": {
                        "description": "Entrada del blog no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Blog"
                ],
                "summary": "Actualizar una entrada del blog por ID",
                "operationId": "d2f1b0701a35fa9fe0a348cf31a0218b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la entrada del blog",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Entrada del blog actualizada exitosamente"
                    },
                    "404": {
                        "description": "Entrada del blog no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Blog"
                ],
                "summary": "Eliminar una entrada del blog por ID",
                "operationId": "d108c8d131119e4b106b1e457c59ec1a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la entrada del blog",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entrada del blog eliminada exitosamente"
                    },
                    "404": {
                        "description": "Entrada del blog no encontrada"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BlogPosts": {
                "title": "BlogPosts",
                "description": "Modelo de entradas de blog",
                "properties": {
                    "id": {
                        "description": "ID de la entrada del blog",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "description": "Título de la entrada del blog",
                        "type": "string"
                    },
                    "content": {
                        "description": "Contenido de la entrada del blog",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación de la entrada del blog",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización de la entrada del blog",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BlogPosts"
                }
            }
        }
    },
    "tags": [
        {
            "name": "Blog",
            "description": "API para gestionar las entradas del blog"
        }
    ]
}